Updated openstack/openstack Project: openstack-infra/storyboard-webclient  cdea7d397f1c367b4b8fb7ed73dc08ed6286f6dc
Make it easier to find 'add task' button

The button was in the top-right of the tasks list, which meant new
users could miss it, or worse, confuse it with the 'add' (tag) button.
This moves the button to the bottom left of the task list, which is
the first place a user is likely to look when they find a task is
missing, and want to add one.

Change-Id: I48c93bc272570a041e7c4437ff82fbd0611d4a04