Updated openstack/openstack Project: openstack/nova  5907f5fffa9a161e720ed0a0d2207832677c241a
libvirt: release serial console ports when destroying guests

The part of code responsible to release serial ports does not cover
all the cases. This commit moves that part inside the method _destroy
which is responsible for destroying the guest from a libvirt
perspective which is when we want to release ports.

Change-Id: I24f2a1fa3b7ef3f2a3196f3689e5f2ba7d49bd87
Closes-Bug: #1489853