Updated openstack/openstack Project: openstack/glance-specs  d00cddc3179aba0be5e67bd5112e09328227d966
Move spec for removing downgrades to right place

The spec for removing downgrades has been approved. The folders for the
specs have been changed in newton and current location of the spec
doens't appear on the correct specs documentation page [1]; rather shows
up on incorrect place [2].

This patch merely moves the spec in the appropriate location.

[1] http://specs.openstack.org/openstack/glance-specs/specs/newton/index.html
[2] http://specs.openstack.org/openstack/glance-specs/

Change-Id: Iaa9a764e1fb372608e59c41db8173f2a848bf697