Updated openstack/openstack Project: openstack-infra/project-config  0d1f3927d46b2482e62ebad4a8457be77cced9ab
Add installguide-cookiecutter repo

Add a new repository to make it easy for new projects to write Install
Guides as discussed in the Newton Install Guide design session.

This repo is owned by the docs team. We will start with an empty
repository without any jobs setup for this.

implements bp project-specific-installguides

Needed-By: I7714905263913ac4ad6ce76fca86c57be367f531
Change-Id: Id3db386827b0b807ba773e972b050db98f28dbe9