Updated openstack/openstack Project: openstack/nova  b4db9329369898eb70e85db2101a211649cd8ba3
api-ref: method verification and fixes for servers.inc

Update the response codes for the methods based on what is in the
code. This follows the pattern of updating these error codes when
doing method verification that jichen has been doing.

Some notes are left for the body_verification phase later, as that is
going to take a while for this file.

Part of bp:api-ref-in-rst

Change-Id: I8d8bd09f72572caed1a45fa23f95a15d330c2f26