Updated openstack/openstack Project: openstack/swift  7429da0ddbc473718a4637c70e29b904435a3092
SwiftLogFormatter will log transaction IDs on INFO level

Previously SwiftLogFormatter would make two checks. One to see if the
transaction id was already in the message field and another check to
make sure the log level wasn't set to info. If either of these was
true, then it would not log the transaction ID in the transaction ID
field.

This commit removes the check for the info log. Now transaction IDs
will be recorded in all cases that have them.

Change-Id: Ic06538ab55a75d298169ae1745671573ee9c09e8
Closes-Bug: #1504344