Updated openstack/openstack Project: openstack/cinder  2a565c652ba8a29558455dddf960e4b44f6493fd
Change OVOs project_id and user_id to StringField

The DB models define project_id and user_id as Column String fields of
255 characters because historically they weren't wasn't necessarily
UUIDs, but when we created the Versioned Object counterparts of the ORM<
models we forced them to be UUIDs, which will create problems for any
deployment where any of them is not using UUIDs.

We know cases where this is a problem for the project_id, moreover now
that Oslo Versioned Objects library is going to enforce the use of UUIDs
in UUIDFields.

This patch changes those fields to make them of StringField type as they
should have been all along.

This change has no implications on the serialization, so we don't need
to bump versions, we can simply change the hash in the test.

Change-Id: Id94ff7777260cc443f838aa3a8399582b2d00fc2