Updated openstack/openstack Project: openstack/horizon  bd8b52cbcdfed2b8e67c297d0fa431b625ac4406
Simplify "Instance Count" verbiage

When launching an instance, having the field "Instance Count" with the
help message of "Number of instances to launch" does not seem very
useful from a users perspective.  Simply naming the field "Number of
Instances" clarifies the purpose of this field and removes the need for
any help message.

Change-Id: If0b0e088bf4d1ba44fd4890c77275eaf182964eb