Updated openstack/openstack Project: openstack/nova  4f8761682a656456ca90dfdf5cd7e88f107e246a
api-ref: add url parameter to expand all sections

When attempting to link to a specific element that is inside of a
collaspable section, the anchors don't work because they are not
visible. A relatively low complexity work around is to add an
?expand_all parameter, which allows one to specify the url as a page
that should be fully expanded. Then any deep linking will take people
to the right place.

Part of bp:api-ref-in-rst

Change-Id: Iae7bc4b2cee9ef76323877d502035e878d0d7182