Updated openstack/openstack Project: openstack/nova  cb0740aa2dcdbe5b571b2c3ed2552d9323bd64cc
api-ref: update parameter validation on servers

This updates all the v2.1 parameters for servers, as well as some of
the microversions added parameters. It makes us more accurate than the
existing api-site.

This specificially *does not* include the additional response
parameters for microversions > 2.1 (tags, ext attrs for ec2, etc).

Part of bp:api-ref-in-rst

Change-Id: Ic03c0e6486a300c3cd724b798850dedeae014faf