Updated openstack/openstack Project: openstack/nova  421c52d9d341b07d850c21e0e702a008a8e1d3b7
api-ref: complete verification for diagnostics.inc

This completes the verification for the diagnostics API. As the
response is still dumping out raw hypervisor stats (for compatibility
https://github.com/openstack/nova/blob/7529b94f4a0be17fb3988362619f83b404ccad86/nova/api/openstack/compute/server_diagnostics.py#L38-L44)
the best we can say about the response is it's hypervisor specific and
don't use it.

In future we should use the standardized format that was defined
during the v3 effort.

Part of bp:api-ref-in-rst

Change-Id: Id2cca6a2f79542bec17a34aa3d31924e8ed7b9e5