Updated openstack/openstack Project: openstack/oslo.versionedobjects  72dad2600488d4c49c34233db27604ed2bdcd675
Fix issue with coercing valid_values to a tuple

A recent patch exposed valid_values via property, and cocerced the
result to a tuple to convey the immutable nature of them. This was
a good thought, but unfortunately breaks __repr__ stability, which
we use in the fixture to calculate the object hashes.

We could make the __repr__ dig into the raw types on the field
instead of using the property, but this is probably not the best and
smallest change at this point. If we can get everyone to start using
tuples for their value_values list, then we could land a change to
coerce and enforce that. Right now, we just need to fix the
unexpected regression.

Change-Id: I09b78e7f816fc26c69a3e9435c9d92d5acb5821b