Updated openstack/openstack Project: openstack-infra/project-config  c83601c267b1784c823fc8e73e3d71743c6580cd
Use neutron-vpnaas grenade plugin in grenade jobs

Now that neutron-vpnaas is a plugin, it needs an explicit
grenade plugin to shutdown the agent for the upgrade. The
devstack neutron code used to handle this but it was removed
in Ie5fc0d2a45c1b564f98c69ec9ea6fbdeeb465d32.

Change-Id: Iaaf02b51dff631758dd908b605f21ff1663fd858
Closes-Bug: #1579235