Updated openstack/openstack Project: openstack-dev/devstack  37f6ea619e64641f80aa59029a71b91bccf2b00a
Export the 'short_source' function & don't keep PS4 in sudo

Export the 'short_source' function so that it will be present in the
environment for child shell scripts. Do this because we are passing PS4
to the child shell scripts and it is using 'short_source'

Don't do an 'env_keep' in the sudoers file for PS4, since it is
difficult to also pass along the 'short_source' function.

Change-Id: I9781010d6eb336d02939c7fd47f18bedeae5ccc6
Closes-Bug: #1563443