Updated openstack/openstack Project: openstack/barbican  66418ecdc7cbaa3fa6405d1f9eb409f8acdfe3e7
Fix creation of notification server

The parameters that the NotificationDispatcher receives changed
recently, and the targets (with its priorities) and requeuing now
depend on the server/listener. So this patch makes the change to
remove those parameters from the creation of that object, and replace
the MessageHandlingServer for a NotificationServer that can actually
handle these things.

Change-Id: Ifc9cfcbd7e80cdd451a79c579294873cbd220c39
Closes-Bug: #1580461