Updated openstack/openstack Project: openstack/governance  9a65dfec3ebe51852f5cfd9030f94016d8344340
Change OpenStack-Ansible release model to cycle-trailing

OpenStack-Ansible (OSA) has been releasing on a cycle-based model
since its inception, but none of the existing models really
fit properly.

The release-trailing model was recently created and provides
the opportunity for OSA to formally implement a cycle-based
release model while still accommodating the need to sometimes
lag a little behind the service projects. This is essential for
deployment projects as they consume the service projects.

This patch changes the OSA release model to cycle-trailing to
properly reflect the model that's always been used and to allow
OSA to properly fit within OpenStack's governance framework.

Change-Id: I5375caefa078a8b34c8b2c24a9fcb6e7ea90b751