Updated openstack/openstack Project: openstack/governance  37dfcd7123405fc1397ab40f3a12841ec3771f75
Adds networking-hyperv to winstackers governance

Commit 846a52873 introduced a selection criterion for Neutron projects
whereby no interface to proprietary system is vouched for by the Neutron
core team.

Because of this, networking-hyperv was removed from the neutron governance
in the commit 396c0df16.

This commit adds networking-hyperv to the winstackers governance. This
prjects fits the winstackers mission, as it is a Neutron ML2 Agent
needed for binding neutron ports to Hyper-V compute nodes.

networking-hyperv was added back in Kilo as a result of the Neutron
vendor decomposition.

Change-Id: Idf4889c9e81e5bbe10ddfdd03fa7e47f4442e6a9