Updated openstack/openstack Project: openstack-infra/project-config  6d1618d7246a3ec508220aaf450c69860ddb65af
Add new project Higgins to openstack repo

Higgins is a Container Management Service for OpenStack to make
container available as first class resources in OpenStack. It aims
to provide an unified API or launching and managing containers
backed by different container technologies.

Magnum is the original OpenStack Container Service, but this is
going to change [1]. In the future, Magnum will focus on deploying
and managing Container Orchestraction Engines (COEs), and the job of
managing container life-cycle will go into Higgins, which is
independent of Magnum.

Higgins is not an official OpenStack project yet, but we will work
hard to make it become an official project.

[1] https://review.openstack.org/#/c/311476/

Change-Id: Ibfcafdc3608299563704708af4cac1ee737c5311