Updated openstack/openstack Project: openstack/django_openstack_auth  dc847867de76830b2a58735e5e1a3c8d3a935d34
When calculating session_time, use the actual token life

This calculation uses the 'token_life' var which is a
datetime.timedelta object.  timedelta.seconds gets us just the
'seconds' component of the object, truncating away any days, hours,
or weeks that might be included in the object.

What we want here is the total time in seconds, which is
total_seconds().

Closes-Bug: #1562452
Change-Id: I6a947abb891e1d34e1cf3aea53b345e0a804bacf