Updated openstack/openstack Project: openstack-infra/project-config  43ce150377b06ab97bfbebe99263cab114353a35
Add js-generator-openstack to OpenStack-Infra

This repository is a project template for the Yeoman project
scaffolding and maintenance framework, which is the JavaScript
equivalent of python's cookiecutter. Its purpose is to handle the
question of "How does OpenStack do JavaScript".

The charter of this project is to roll up several project
maintenance concerns which, in our Python projects, are spread
between cookiecutter, requirement management, and infra. The justification
is both usability and usage. Usability, so that creating and/or updating
a new JavaScript project is simple, and to create 'one place' from which
we publish changes. Usage, because placing more concerns into one project
will be easier to manage for today's limited number of JavaScript devs.

The goal is that at any point, an engineer may run `yo openstack` at the
root level of an existing (or new) javascript project, and expect their
project to be updated to the latest requirements, tools, and expected
minimum gate criteria.

Because this project includes the concerns of several teams, it has been
proposed under the OpenStack-Infra program. The reason is that of the
three major concerns - requirements, CTI, and templating - the first is
slated to be moved under the infra banner now that it's stable, and the
second is already under the infra banner.

Change-Id: Ia60a1445ad2fa48a9667b00bce8a0f67a87f5f0c
Needed-by: Ia4c1e1b210b6bf9c973f4f83ca3856d83ca28824