Updated openstack/openstack Project: openstack/neutron  b6c58e0f5219bc4ce8821b7fac007bef50ec3a87
agentnotifiers: retried notifications refactor

This patch introduces retry(func, max_attempts) method
which wraps the original func in such a way,
that if execution results in MessagingException,
the given function will be retried for max_attempts
till success or MessagingException is raised.
The function is in the utils module and can be reused
by different agentnotifiers if necessary.

Change-Id: I0d0c17e500e44c1a17438c29a0e76a9ef00872e8