Updated openstack/openstack Project: openstack/nova  7ec1472380f485dda33973e1f24f4cf7748902b5
Added new exception to handle CinderClientException

When Cinder client exception is thrown in initialize_connection
(which is called during prechecks for Live migration),
instance is moved to ERROR state. It is not sensible
to move the instance to ERROR state when the precheck fails.
Adding new exception changes this behavior. Instance will be reset
to Active state when the precheck fails.

Closes bug: #1544744

Change-Id: I7a5fcc070ff53086f37417f12e2b9f383e220747