Updated openstack/openstack Project: openstack/keystonemiddleware  f1fbf82aac5fd06d2bb4d8c4d83f644f50802333
Handle cache invalidate outside cache object

Move the logic that handles invalid cache entries out of the cache
object itself. This makes the cache interface more normal and will make
it easier to convert to oslo.cache later.

Rename the store method to set because this is the interface that
oslo.cache uses.

Related-Bug: #1523375
Change-Id: Ia7b11a69f14e17a9d7a36c009d9a965be16509d0