Updated openstack/openstack Project: openstack/neutron  b84d33ec6ee1c6ceeb36ae798b32d258554ee23e
Avoid L3 agent termination without server

This patch fixes the termination of L3 agent if the neutron
server is not available. It was trying to query service plugins
enabled on the server, which is used to determine if FWaaS is
enabled or not. When L3 agent is unable to contact neutron
server to retrieve plugins list, it will be retried infinitely
with an increasing timeout between each try.

The timeout logic is implemented as a separate patch
https://review.openstack.org/#/c/280595/

There's been an earlier fix which alleviates this problem to
some extent with bug #1368795. This patch modifies the logic
to keep retrying forever, instead of terminating the agent.
It also adds a more clear message to the user in logs.

Change-Id: If0b136817c7bca8fcd2b660ba88ae69c3f341764
Closes-Bug: #1569921