Updated openstack/openstack Project: openstack/nova  92aaa4d40bcd291e9d354e16c9986a057ee42c73
db: retry instance_info_cache_update() on deadlock

If a Galera cluster is used in multi-writer mode it's possible, that
instance_info_cache_update() will be executed concurrently on two
different MySQL hosts for the very same row, which causes a deadlock
exception for one of the callers due to how Galera works internally.

This can affect operations like association or disassociation of
floating IPs, which will fail, if instance_info_cache_update() does
not handle deadlocks gracefully, i.e. is not retried.

Closes-Bug: #1567336

Change-Id: Ib5abffd94d2480dfbcc8b6cca7b1c73ce39e7d10