Updated openstack/openstack Project: openstack/requirements  dbdcf7b1e7fd098199bc5bf224efc78974e4df2e
Block dnspython 1.13.0

DNS Python broke the dns.zone.from_text() method

It now requires an extra parameter (the origin of the
zone), and we do not have that information when we
call into dnspython.

Previously the origin was taken from the $ORIGIN
statement in the zone file, and we relyed on this
to parse the name of the zone when we run an
import.

Change-Id: Ib3849cd5c0f4908c0b19787ed84891d7a49f63e1
Related-Bug: #1580571