Updated openstack/openstack Project: openstack/horizon  2de6ed815d3480087e9a1cbbc5aceec5849717b9
Update tox.ini for pot files not in tree

The translation import fails since we do not store pot files in tree
anymore. Replace the "git checkout" after testing message extracting
with a simple rm. Note that we cannot use a wildcard in tox.ini for the
rm, instead give full names.

This is a followup to Ief0861e01ff758120001f4c8dfdbf9afadf09f5b.

Change-Id: I0fc59a444cfca9ffdf7e7eaa73ba321295c2aa2e