Updated openstack/openstack Project: openstack/neutron  ad66246742122d0ab388b665a10ed42c57fc94a0
Enable IPAM drivers to allocate from more than one subnet

Calling in to the IPAM driver independently for each subnet on a
network can be costly, especially if the IPAM system is external.
This change opens the possibility of optimizing the operation of
allocating one IP address from a list of candidate subnets by allowing
drivers to subclass the default implementation.

Change-Id: I07b32752e1e2e84641dd6504a93cfef95ae368c5
Partially-Implements: blueprint bp/routed-networks