Updated openstack/openstack Project: openstack-dev/grenade  ae1837e18da9da155bb9d2bac2cc4dbfd4c17a2f
create early_create phase

For projects that need to setup custom networking we need an
early_create phase that has only gotten the base neutron network setup
which they can then work with and make changes as appropriate.

Create this new hook point, and move the neutron create to this early
phase.

Change-Id: Ifb1903f8d3cb11bc4db33f8acdea342e6373f864