Updated openstack/openstack Project: openstack/nova  677594480ecb9c093a3d81ae49dead120798a5c4
Fix race condition for live-migration-force-complete

Live-migration monitor is single source of truth for
actual migration state. This patch introduce mechanics
that allows to schedule action, e.g. force-complete on
active migration, and execute it in the right time.
It fix issue that allow to leave active instance paused
after live-migration.

Closes-bug: #1580609

Change-Id: Iadd8f57949b99a893b7c2c3a19084a21ae6290f9