Updated openstack/openstack Project: openstack/keystone-specs  d33d6521892018be9cb7a0dc2106906c9be9f68c
Improve example of project acting as a domain

The current project example in the project entity section was,
in fact, invalid - it showed a project acting as a domain that
was not at the top of the tree. This patch corrects this and adds
correct examples of a project acting as a domain and a top level
regular project.

Change-Id: I2b0519f3c0ea274898ffd2ed6cad24c6b4352940