Updated openstack/openstack Project: openstack/cinder  73868becf375d4409e6c7ee9cf4130ab1303294c
Fix uncaught NotFound exceptions

There are a couple of places in the code where we were not catching
NotFound exceptions which resulted in getting the exception logged
incorrectly in the logs and the error message being returned by the WSGI
layer being slightly off.

Cases where this was happening are volume encryption metadata and volume
transfer.

From the user perspective the error was very similar to what it was
expected, but it was preceeded by the name of the Cinder exception (ex:
VolumeNotFound: ).

This patch catches all NotFound exception at the WSGI level and properly
converts them, that way we prevent the exception to propagate to the
middleware fault level and generate a log exception and return slightly
different error.

Change-Id: I8971afedd68854d2840ab551b420ca083e8ee0d6