Updated openstack/openstack Project: openstack/nova  355fa5307fb436f161693e09da21da9deaa8bea0
Complete method verification of os-networks

Verified the API document with source for methods available and
response codes used.  Recorded HTTP methods to match wiki.

Marked the associate/disassociate sections as deprecated (as they are
nova-net only), and added some warnings in the docs about their use.

Part of bp:api-ref-in-rst

Change-Id: I63c56a546fe3a97eaa16ce41999593eaf650b538