Updated openstack/openstack Project: openstack/neutron  b25f6448d530b51eea4a36b3d087515e8495f53f
Ensure ML2's create/update_port methods not in transaction

This adds a check to the ML2 create and update port methods which
are called by other services to manipulate ports. This check prevents
them from passing in a context that is already part of an ongoing DB
session because we do not want DB rollbacks to be allowed after the
ML2 plugin calls postcommit methods on drivers.

This also adds a temporary hack to set an attribute on the context
to skip this check to accomadate two L3 code-paths and a subnet
code-path that have port manipulations entangled in transactions.
This attribute will ultimately be removed once these paths are
refactored.

Closes-Bug: #1540844
Change-Id: I5aa099c2264636336ab0b76c0826b506e2dc44b6