Updated openstack/openstack Project: openstack-infra/project-config  8398ad297b569cd7323d85895dec9f79d0a77e04
Import ldappool into gerrit and setup project

The OpenStack Keystone Team (a subset) is adopting the ldappool code
base from the original author so that we can update/maintain and
python-3 enable the library. I have discussed this with the original
owner and the project is currently unmaintained and we are welcome
to bring it in (since Keystone relies on it).

Change-Id: Id75d304b3e4af98e64d97db29a3aa6c3ca55354e