Updated openstack/openstack Project: openstack-infra/shade  7dd138c8f4c8b8761b0b50d3ce9c7c4f770f8c61
Add floating IPs to server dict ourselves

The nova/neutron interaction as it relates to updating the network info
cache in nova apparently does not believe in an object reality and
instead prefers to think of things as unrealized quantum superpositions.
While those are great for particle physics, they make getting the
address of the server hard. It therefore follows that we should
undertake the burden of providing consistent and objective information
about the floating ip of the server ourselves, regardless of the sadness
this adds to the world.

Change-Id: I75ba0318c9a5a011b5a530d86ace275f19511f1f