Updated openstack/openstack Project: openstack/horizon  42e3c6884e74c0b7320d7ddc90d2422d4110e141
Add Line Charts to the Theme Preview Page

Line Charts were missing.  Lets add them to make
it easier when we need to make them themable or
debug them.

There is currently a problem with the charts hover over
label.  It looks like this is an existing bug, and will
not be fixed in this review.  Fix will come soon when
the charts are made themable.

The examples on the d3linechart file were wrong and did
not work with the code.

Partially-Implements blueprint bootstrap-theme-preview
Change-Id: Icf278c6b2e4605a4e867e20080a59403683fae7f