Updated openstack/openstack Project: openstack/requirements  4aeff3ffe394015af9b0b0197f8fe292ed833880
Upgrade PyKMIP to 0.5.0

The 0.5.0 release of PyKMIP includes a new PyKMIP server that can be
used as a key manager for testing purposes. This feature is required
for the proposed Barbican KMIP gate check.

Updating global-requirements to this version will allow a gate check
that can test the Barbican KMIP plugin with an actual KMIP server.

Change-Id: Ifda13a84607bb199b794dc24f5dbba0ee8108dbf