Updated openstack/openstack Project: openstack/nova  d170671eaa609f12fa2a0060a482500b70e95798
Remove mox in test_volume_snapshot_create_outer_success

This replaces usage of mox with mock and also adds mocking of the
FixedIntervalLoopingCall in the test as there was a race on whether
or not _wait_for_snapshot and subsequently _volume_api.get_snapshot
would be called by the time the call assertion of get_snapshot
was made.

This was uncovered when https://review.openstack.org/#/c/315241/
replaced mox with mock in completely different tests and
test_volume_snapshot_create_outer_success began failing when run
in parallel with other tests. When run alone, it passed.

Part of blueprint remove-mox-newton

Change-Id: I4952f5cb3a3765ea0dee4a2a38483b6434d65ce6