Updated openstack/openstack Project: openstack/keystone  0f579aa362f021b17f7c2931ffca309e395bd51f
Move the assignment abstract base class out of core

This patch moves the assignment abstract base class out of core and into
backends/base.py

This removes dependencies where backend code references code in the
core. The reasoning being that the core should know about the backend
interface, but the backends should not know anything about the core
(separation of concerns). And part of the risk here is a potential for
circular dependencies.

Partial-Bug: #1563101

Change-Id: Ie507b61044ad32ecaac26901c1bcd0a0c17bd381