Updated openstack/openstack Project: openstack/zaqar  3f15c0bee148721217bcf54556bf9e9042544746
Move to falcon middleware

Instead of using the now removed hook feature of falcon, let's use
the middleware feature.

NOTE: Though the test is changed. But actually it's not correct since
the unicode test with a running zaqar server will return 404 instead
of 400. But it's out of the scope of this fix. Will fix it in a
following patch.

Co-Authored-By: Fei Long Wang <flwang@catalyst.net.nz>

Closes-Bug: #1581189
Change-Id: I7e8e97b73ec3e245580f07e62442caf0395d5745