Updated openstack/openstack Project: openstack/nova  de3eb4913f045c2c5827bb320230198e51620321
Eager load keypairs in instance metadata

In the review for e83842b80b73c451f78a4bb9e7bd5dfcebdefcab it was
pointed out that keypairs should be added to the eager load list for
get_metadata_for_instance_id. Due to a lack of a proper context in a
thread in the ironic driver, this actually totally broke ironic.
The context problem is being fixed elsewhere, but this should be
eager-loaded to minimize DB queries anyway.

Change-Id: I67b4665bdebd45c66983c6a556d20f367fa7fbdf
Closes-Bug: #1580987