Updated openstack/openstack Project: openstack/cinder  1246fadb38a88b01e9a6523d8ba5d2dc00f8485c
Change provider_id to StringField

Volume and Snapshot objects had provider_id defined as a
UUIDField. This field is for drivers to stash away useful
identifying info that may or may not be a UUID. There are
existing drivers that just store a string here.

This is a compatible change, so just switching over the
field from UUID to String to get rid of oslo warning
messages when non-UUID values are stored in this field.

Change-Id: I3ab4080b8139dab5ca336f8988093025b0b6dfbe