Updated openstack/openstack Project: openstack/diskimage-builder  c3cb37577101d5244f1208b9bda89b860f18c3b1
Split YAML & JSON parsing

It turns out that invalid JSON can be valid YAML ... thus if you mess
up a pkg-map file that still works as a YAML file dib-lint will let it
pass, but when pkg-map later tries to open it as a JSON file, it
fails.

Parse each type separately to catch these problems.

Change-Id: Ib3985e7d1599ed6bf3b7a73b786a53177b71fae0